+++ src/ts/components/registry.ts
@@ -59,7 +59,7 @@ export const AUTO_INIT_COMPONENTS = {
   Tooltip: { component: Components.Tooltip, selector: ".tooltipped" },
   FloatingActionButton: {
     component: Components.FloatingActionButton,
-    selector: ".fab, .fixed-action-btn, .fab-menu",
+    selector: ".fab-menu",
   },
 };